Move tray password dialogs off Linux builds.
golangci unused fails on linux/freebsd because those helpers are only called from the Windows/Darwin tray.
